+++ teste-gigalink/src/main/java/br/com/jfcardoso/testegigalink/entities/Pedido.java
@@ -3,15 +3,16 @@ package br.com.jfcardoso.testegigalink.entities;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
 @Builder
 public class Pedido {
-    private int id;
+    private Long id;
     private Date dataHora;
     private String notaFiscal;
-    private double valorFrete;
-    private double desconto;
-    private double valorTotal;
+    private BigDecimal valorFrete;
+    private BigDecimal desconto;
+    private BigDecimal valorTotal;
 }
